requirements.txt
flask
flask-cors
numpy
pandas
tensorflow-cpu
scikit-learn
openpyxl
gunicorn
h5py